+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/http/ClientCertKeyManager.kt
@@ -2,17 +2,12 @@ package ch.rmy.android.http_shortcuts.http
 
 import android.content.Context
 import android.security.KeyChain
-import okhttp3.OkHttpClient
 import java.net.Socket
-import java.security.KeyStore
 import java.security.Principal
 import java.security.PrivateKey
 import java.security.cert.CertificateException
 import java.security.cert.X509Certificate
-import javax.net.ssl.SSLContext
-import javax.net.ssl.TrustManagerFactory
 import javax.net.ssl.X509KeyManager
-import javax.net.ssl.X509TrustManager
 
 class ClientCertKeyManager(
     private val alias: String,
@@ -41,20 +36,7 @@ class ClientCertKeyManager(
     }
 
     companion object {
-        fun applyToOkHttpClient(builder: OkHttpClient.Builder, context: Context, alias: String): OkHttpClient.Builder {
-            val sslContext = SSLContext.getInstance("TLS")
-            sslContext.init(arrayOf(getClientCertKeyManager(context, alias)), null, null)
-            return builder.sslSocketFactory(sslContext.socketFactory, getDefaultTrustManager())
-        }
-
-        private fun getDefaultTrustManager(): X509TrustManager {
-            val trustManagerFactory = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm())
-            trustManagerFactory.init(null as KeyStore?)
-            return (trustManagerFactory.trustManagers.singleOrNull() as? X509TrustManager)
-                ?: throw IllegalStateException("Failed to get default trust manager")
-        }
-
-        private fun getClientCertKeyManager(context: Context, alias: String): ClientCertKeyManager {
+        fun getClientCertKeyManager(context: Context, alias: String): ClientCertKeyManager {
             val certChain = KeyChain.getCertificateChain(context, alias)
             val privateKey = KeyChain.getPrivateKey(context, alias)
             if (certChain == null || privateKey == null) {